Fikser sånn at alle tester kjører
* surefire tar per default ikke med tester som slutter på *Enhetstest
* Renamer test som mangler test suffix
* Splitter ut test fra BeregnetAndelDsl
